+++ modules/src/ics02_client/client_state.rs
@@ -4,6 +4,7 @@ use std::time::Duration;
 
 use prost_types::Any;
 use serde::{Deserialize, Serialize};
+use tendermint::trust_threshold::TrustThresholdFraction;
 use tendermint_proto::Protobuf;
 
 use ibc_proto::ibc::core::client::v1::IdentifiedClientState;
@@ -55,6 +56,15 @@ impl AnyClientState {
 
             #[cfg(any(test, feature = "mocks"))]
             Self::Mock(mock_state) => mock_state.latest_height(),
+        }
+    }
+
+    pub fn trust_threshold(&self) -> Option<TrustThresholdFraction> {
+        match self {
+            AnyClientState::Tendermint(state) => Some(state.trust_level),
+
+            #[cfg(any(test, feature = "mocks"))]
+            AnyClientState::Mock(_) => None,
         }
     }