+++ relayer/src/supervisor.rs
@@ -7,12 +7,11 @@ use std::{
 use anomaly::BoxError;
 use crossbeam_channel::{Receiver, Sender};
 use itertools::Itertools;
-
-use tracing::{debug, error, info};
+use tracing::{debug, error, info, trace, warn};
 
 use ibc::{
     events::IbcEvent,
-    ics24_host::identifier::{ChainId, ChannelId},
+    ics24_host::identifier::{ChainId, ChannelId, PortId},
     Height,
 };
 
@@ -28,9 +27,12 @@ use crate::{
     worker::{WorkerMap, WorkerMsg},
 };
 
+pub mod client_state_filter;
 mod error;
+
+use client_state_filter::{FilterPolicy, Permission};
+
 pub use error::Error;
-use ibc::ics24_host::identifier::PortId;
 
 pub mod dump_state;
 use dump_state::SupervisorState;
@@ -59,6 +61,7 @@ pub struct Supervisor {
 
     cmd_rx: Receiver<SupervisorCmd>,
     worker_msg_rx: Receiver<WorkerMsg>,
+    client_state_filter: FilterPolicy,
 
     #[allow(dead_code)]
     telemetry: Telemetry,
@@ -70,6 +73,7 @@ impl Supervisor {
         let registry = Registry::new(config.clone());
         let (worker_msg_tx, worker_msg_rx) = crossbeam_channel::unbounded();
         let workers = WorkerMap::new(worker_msg_tx, telemetry.clone());
+        let client_state_filter = FilterPolicy::default();
 
         let (cmd_tx, cmd_rx) = crossbeam_channel::unbounded();
 
@@ -79,49 +83,93 @@ impl Supervisor {
             workers,
             cmd_rx,
             worker_msg_rx,
+            client_state_filter,
             telemetry,
         };
 
         (supervisor, cmd_tx)
     }
 
-    fn relay_on_channel(
+    /// Returns `true` if the relayer should filter based on
+    /// client state attributes, e.g., trust threshold.
+    /// Returns `false` otherwise.
+    fn client_filter_enabled(&self) -> bool {
+        // Currently just a wrapper over the global filter.
+        self.config.read().expect("poisoned lock").global.filter
+    }
+
+    /// Returns `true` if the relayer should filter based on
+    /// channel identifiers.
+    /// Returns `false` otherwise.
+    fn channel_filter_enabled(&self) -> bool {
+        self.config.read().expect("poisoned lock").global.filter
+    }
+
+    fn relay_packets_on_channel(
         &self,
         chain_id: &ChainId,
         port_id: &PortId,
         channel_id: &ChannelId,
     ) -> bool {
-        !self.config.read().expect("poisoned lock").global.filter
-            || self
-                .config
-                .read()
-                .expect("poisoned lock")
-                .find_chain(chain_id)
-                .map_or_else(
-                    || false,
-                    |chain_config| {
-                        chain_config
-                            .filters
-                            .channels
-                            .contains(&(port_id.clone(), channel_id.clone()))
-                    },
-                )
-    }
-
-    fn relay_on_object(&self, chain_id: &ChainId, object: &Object) -> bool {
-        if !self.config.read().expect("poisoned lock").global.filter {
+        // If filtering is disabled, then relay all channels
+        if !self.channel_filter_enabled() {
             return true;
         }
 
-        match object {
-            Object::Client(_) => true,
-            Object::Channel(c) => {
-                self.relay_on_channel(chain_id, c.src_port_id(), c.src_channel_id())
+        self.config
+            .read()
+            .expect("poisoned lock")
+            .packets_on_channel_allowed(chain_id, port_id, channel_id)
+    }
+
+    fn relay_on_object(&mut self, chain_id: &ChainId, object: &Object) -> bool {
+        // No filter is enabled, bail fast.
+        if !self.channel_filter_enabled() && !self.client_filter_enabled() {
+            return true;
+        }
+
+        // First, apply the channel filter
+        if let Object::Packet(u) = object {
+            if !self.relay_packets_on_channel(chain_id, u.src_port_id(), u.src_channel_id()) {
+                return false;
             }
-            Object::Packet(u) => {
-                self.relay_on_channel(chain_id, u.src_port_id(), &u.src_channel_id())
+        }
+
+        // Second, apply the client filter
+        let client_filter_outcome = match object {
+            Object::Client(client) => self
+                .client_state_filter
+                .control_client_object(&mut self.registry, client),
+            Object::Connection(conn) => self
+                .client_state_filter
+                .control_conn_object(&mut self.registry, conn),
+            Object::Channel(chan) => self
+                .client_state_filter
+                .control_chan_object(&mut self.registry, chan),
+            Object::Packet(u) => self
+                .client_state_filter
+                .control_packet_object(&mut self.registry, u),
+        };
+
+        match client_filter_outcome {
+            Ok(Permission::Allow) => true,
+            Ok(Permission::Deny) => {
+                warn!(
+                    "client filter denies relaying on object {}",
+                    object.short_name()
+                );
+
+                false
             }
-            Object::Connection(_) => true,
+            Err(e) => {
+                warn!(
+                    "denying relaying on object {}, caused by: {}",
+                    object.short_name(),
+                    e
+                );
+
+                false
+            }
         }
     }
 
@@ -266,7 +314,13 @@ impl Supervisor {
 
     /// Create a new `SpawnContext` for spawning workers.
     fn spawn_context(&mut self, mode: SpawnMode) -> SpawnContext<'_> {
-        SpawnContext::new(&self.config, &mut self.registry, &mut self.workers, mode)
+        SpawnContext::new(
+            &self.config,
+            &mut self.registry,
+            &mut self.client_state_filter,
+            &mut self.workers,
+            mode,
+        )
     }
 
     /// Spawn all the workers necessary for the relayer to connect
@@ -515,9 +569,9 @@ impl Supervisor {
 
         for (object, events) in collected.per_object.drain() {
             if !self.relay_on_object(&src_chain.id(), &object) {
-                info!(
+                trace!(
                     "skipping events for '{}'. \
-                    reason: filtering is enabled and channel does not match any enabled channels",
+                    reason: filtering is enabled and channel does not match any allowed channels",
                     object.short_name()
                 );